modules/IoTEdgeCustomMetrics/requirements.txt
azure-iot-device~=2.7.0
prometheus-client
pytz